# [Xamarin.Android.Build.Tasks] #deletebinobj and ResolveLibraryProjectImports