Fixes: http://work.devdiv.io/760829

I've recently discovered a scenario that requires a `Rebuild` to fix...

1. File -> New Xamarin.Forms Project in IDE
2. Do *not* build! Note that the IDE will do a build for Intellisense:
   design-time build in VS Windows and `UpdateAndroidResources` in VS
   Mac
3. Update NuGet packages (they are `<PackageReference/>`)
4. Build

You will likely hit `APT0000` errors and/or Java compilation errors:

    /Users/builder/Library/Developer/Xamarin/jdk/microsoft_dist_openjdk_1.8.0.10/bin/javac -J-Dfile.encoding=UTF8 -d obj/Release/android/bin/classes -classpath /Library/Frameworks/Mono.framework/External/xbuild-frameworks/MonoAndroid/v8.1/mono.android.jar:obj/Release/lp/0/jl/formsviewgroup.jar:obj/Release/lp/10/jl/bin/classes.jar:obj/Release/lp/11/jl/bin/classes.jar:obj/Release/lp/12/jl/bin/classes.jar:obj/Release/lp/13/jl/bin/classes.jar:obj/Release/lp/14/jl/bin/classes.jar:obj/Release/lp/15/jl/bin/classes.jar:obj/Release/lp/16/jl/bin/classes.jar:obj/Release/lp/17/jl/bin/classes.jar:obj/Release/lp/18/jl/bin/classes.jar:obj/Release/lp/2/jl/arch-core-common.jar:obj/Release/lp/20/jl/bin/classes.jar:obj/Release/lp/3/jl/arch-lifecycle-common.jar:obj/Release/lp/4/jl/bin/classes.jar:obj/Release/lp/5/jl/bin/classes.jar:obj/Release/lp/7/jl/bin/classes.jar:obj/Release/lp/8/jl/bin/classes.jar:obj/Release/lp/9/jl/bin/classes.jar:obj/Release/lp/6/jl/support-annotations.jar -bootclasspath /Users/builder/Library/Android/sdk/platforms/android-27/android.jar -encoding UTF-8 "@/var/folders/th/stt5g8t53jzfbz8ryh5_xrsr0000gn/T/tmp7ceebc47.tmp"
        obj/Release/android/src/mono/android/support/v7/widget/RecyclerView_ItemAnimator_ItemAnimatorFinishedListenerImplementor.java:8: error: package android.support.v7.widget.RecyclerView.ItemAnimator does not exist
                android.support.v7.widget.RecyclerView.ItemAnimator.ItemAnimatorFinishedListener

You have to `Rebuild` or #deletebinobj to solve the problem.

After digging in, I started to see problems with
`<ResolveLibraryProjectImports/>` that have just been here a while.

## Files.ExtractAll and timestamps ##

We had a `forceUpdate` flag, which is set to `false` in
`<ResolveLibraryProjectImports/>`:

    if (forceUpdate || entry.ModificationTime > dt) {
        try {
            entry.Extract (destination, fullName, FileMode.Create);
            MonoAndroidHelper.SetWriteable (outfile);
            var utcNow = DateTime.UtcNow;
            File.SetLastWriteTimeUtc (outfile, utcNow);
            File.SetLastAccessTimeUtc (outfile, utcNow);

So assuming the zip file being extracted changes, there are some
problems with what we are doing here:

* By setting `UtcNow` as the timestamp. No file would ever get
  updated... * We made this change in: c1c71d9
* We can't rely on `entry.ModificationTime`, the timestamp in a zip
  file is not reliable at all... It could be `DateTime.MinValue`,
  non-Utc, etc. We can't rely on it...

Basically I can't see how this `forceUpdate=false` setting could be
correct... So I have taken the approach to just remove it.

## ResolveLibraryProjectImports and timestamps ##

In addition to the timestamp logic in `Files.ExtractAll`, we had
similar timestamp logic on assemblies:

    var stamp = new FileInfo (Path.Combine (outdir.FullName, assemblyIdentName + ".stamp"));
    if (stamp.Exists && stamp.LastWriteTimeUtc > new FileInfo (assemblyPath).LastWriteTimeUtc) {
        // ...
        continue;

The problem here, is I don't think we can rely on the timestamps of
these assemblies either!

* `<PackagReference/>` now puts a central NuGet cache in
  `%userprofile%\.nuget`. You could have Xamarin.Forms (or other)
  NuGets cached with very old timestamps.
* Upgrading/downgrading NuGet packages of a cached NuGet could replace
  an assembly with an older timestamp than the stamp file, but we
  wouldn't extract any files!

So the fix I'm thinking here is to use hashes of these files instead
of timestamps. Any file extracted in `<ResolveLibraryProjectImports/>`
will get "re-extracted" if the hash of the file changes. We use the
same stamp files, but store hashes inside them.

## Files.ExtractAll and deleteCallback ##

After these initial set of changes, the original scenario was still
not working... `<ResolveLibraryProjectImports/>` seemed to be leaving
behind old files when NuGet packages were changed.

On incremental builds, I could get things to work properly if I added
a call to:

    Directory.Delete (importsDir, recursive: true);

This basically nuked the `lp` directory on builds where the assembly
changed.

After more digging, it seemed the semantics of `deleteCallback` were
reversed?

    if (!files.Contains (outfile) && !(deleteCallback?.Invoke (outfile) ?? true)) {
        File.Delete (outfile);

An example of `deleteCallback` from `<ResolveLibraryProjectImports/>`:

    deleteCallback: (fileToDelete) => {
        return !files.Contains (fileToDelete);
    }

This seems backwards?

Reviewing the changes in 3f3c45c (Aug 2017), we were deleting the
files here no matter what before `deleteCallback` was added.

I *think* the correct behavior here is to remove the `!`:

    if (!files.Contains (outfile) && (deleteCallback?.Invoke (outfile) ?? true)) {
        File.Delete (outfile);

Now I am seeing `<ResolveLibraryProjectImports/>` delete files. We
don't need a `Directory.Delete ()` call.

The only other place `Files.ExtractAll` is used is in the
`<GetAdditionalResourcesFromAssemblies/>` MSBuild task. Since it does
not pass a callback, I believe the new behavior is correct.

## List of jar files ##

Before we were doing a recursive search for `*.jar` files:

    outdir.EnumerateFiles ("*.jar", SearchOption.AllDirectories)

Since this could traverse a lot of directores, let's add jar files to
the list as we go. It turns out this change is also *necessary*, since
we have fixed `Files.ExtractAll` so it properly deletes files. Without
this change, jar files were getting deleted.

## Tests ##

The `CompileBeforeUpgradingNuGet` test reproduces the original
scenario.

I added a set of tests for `Files.ExtractAll` as well, which helped me
discover most of the bugs there.

One test fix:

`Xamarin.Android.Support.v4.dll` was not even *in use* when updating
Xamarin.Forms. We should be checking
`Xamarin.Android.Support.v7.AppCompat.dll` instead.

## Performance Implications ##

These changes don't seem to have a negative impace on performance:

    Before:
    1st ->      1346 ms  ResolveLibraryProjectImports
    C# change -> 346 ms  ResolveLibraryProjectImports
    After:
    1st ->      1213 ms  ResolveLibraryProjectImports
    C# change -> 366 ms  ResolveLibraryProjectImports

The initial build may be slightly faster. Incremental builds seems to
vary, but roughly be the same.

## Future Work ##

In dd36614, I added a `_CleanIntermediateIfNuGetsChange` target to
combat #deletebinobj problems such as this. I believe the fixes here
will make this target completely unnecessary.

Removing `_CleanIntermediateIfNuGetsChange` will improve build
performance when changing NuGet packages, since right now:

* A new NuGet package will trigger `_CleanIntermediateIfNuGetsChange`
* A "light clean" of `obj` occurs
* Basically everything rebuilds