@jonathanpeppers 2019-02-01
So there could be the case:
- You have a `<ProjectReference/>` to a Xamarin.Android library project
- It has lots of assets, resources, etc.
- You make a code change in the Xamarin.Android library

@dellis1972 do you know of a MonoGame sample with lots of `Assets` I could time? I could move their `Assets` directory to a new project to test this.