@jonathanpeppers 2019-02-01
Ok, I have more changes here, let's see if I can get this green. I will need to rewrite the commit message & re-run perf timings.