@dellis1972 2019-02-08
@jonathanpeppers I think this is a step in the right direction in making the builds more stable. So I will review and merge this I think.

What we need to do is try to come up with a method of handling the specific situation where the zip is that large.